refactor: add database repositories
